pyproject.toml
[project]
name = "jsonschema-codegen"
# dynamic = ["version"]
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Hiroki Teranishi", email = "teranishi.hiroki@gmail.com" },
]
dependencies = [
  "jinja2",
  "pyyaml",
  "referencing>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.13",
  "pytest>=8.3",
  "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]
ignore = ["E501", "B024", "B905", "B019"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
  "yaml.*",
]
ignore_missing_imports = true
